+++ src/Symfony/Component/Workflow/Marking.php
@@ -30,22 +30,22 @@ class Marking
         }
     }
 
-    public function mark(string $place)
+    public function mark(string $place): void
     {
         $this->places[$place] = 1;
     }
 
-    public function unmark(string $place)
+    public function unmark(string $place): void
     {
         unset($this->places[$place]);
     }
 
-    public function has(string $place)
+    public function has(string $place): bool
     {
         return isset($this->places[$place]);
     }
 
-    public function getPlaces()
+    public function getPlaces(): array
     {
         return $this->places;
     }