+++ src/Symfony/Component/Workflow/Validator/StateMachineValidator.php
@@ -19,7 +19,7 @@ use Symfony\Component\Workflow\Exception\InvalidDefinitionException;
  */
 class StateMachineValidator implements DefinitionValidatorInterface
 {
-    public function validate(Definition $definition, string $name)
+    public function validate(Definition $definition, string $name): void
     {
         $transitionFromNames = [];
         foreach ($definition->getTransitions() as $transition) {