+++ src/Symfony/Component/Workflow/Validator/WorkflowValidator.php
@@ -27,7 +27,7 @@ class WorkflowValidator implements DefinitionValidatorInterface
         $this->singlePlace = $singlePlace;
     }
 
-    public function validate(Definition $definition, string $name)
+    public function validate(Definition $definition, string $name): void
     {
         // Make sure all transitions for one place has unique name.
         $places = array_fill_keys($definition->getPlaces(), []);