+++ src/Symfony/Component/Workflow/Workflow.php
@@ -51,7 +51,7 @@ class Workflow implements WorkflowInterface
     /**
      * {@inheritdoc}
      */
-    public function getMarking(object $subject)
+    public function getMarking(object $subject): Marking
     {
         $marking = $this->markingStore->getMarking($subject);
 
@@ -93,7 +93,7 @@ class Workflow implements WorkflowInterface
     /**
      * {@inheritdoc}
      */
-    public function can(object $subject, string $transitionName)
+    public function can(object $subject, string $transitionName): bool
     {
         $transitions = $this->definition->getTransitions();
         $marking = $this->getMarking($subject);
@@ -152,7 +152,7 @@ class Workflow implements WorkflowInterface
     /**
      * {@inheritdoc}
      */
-    public function apply(object $subject, string $transitionName, array $context = [])
+    public function apply(object $subject, string $transitionName, array $context = []): Marking
     {
         $marking = $this->getMarking($subject);
 
@@ -220,7 +220,7 @@ class Workflow implements WorkflowInterface
     /**
      * {@inheritdoc}
      */
-    public function getEnabledTransitions(object $subject)
+    public function getEnabledTransitions(object $subject): iterable
     {
         $enabledTransitions = [];
         $marking = $this->getMarking($subject);
@@ -238,7 +238,7 @@ class Workflow implements WorkflowInterface
     /**
      * {@inheritdoc}
      */
-    public function getName()
+    public function getName(): string
     {
         return $this->name;
     }
@@ -246,7 +246,7 @@ class Workflow implements WorkflowInterface
     /**
      * {@inheritdoc}
      */
-    public function getDefinition()
+    public function getDefinition(): Definition
     {
         return $this->definition;
     }
@@ -254,7 +254,7 @@ class Workflow implements WorkflowInterface
     /**
      * {@inheritdoc}
      */
-    public function getMarkingStore()
+    public function getMarkingStore(): MarkingStoreInterface
     {
         return $this->markingStore;
     }