+++ src/Symfony/Component/Workflow/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+5.2.0
+-----
+
+ * Added function `getEnabledTransition` to easily retrieve a specific transition object
+
 5.1.0
 -----